changed type for some specifiers

+++ base_print.c
@@ -10,7 +10,7 @@
 int octal_print(va_list arg, flag_t *flag, int length)
 {
 	int i = 0, digit = 0, j = 0, k;
-	unsigned int num = va_arg(arg, unsigned int);
+	unsigned long int num = va_arg(arg, unsigned long int);
 
 	num = lengthConverter(num, length);
 	digit = base_converter(num, 8, -1, 0);
@@ -52,7 +52,7 @@ int octal_print(va_list arg, flag_t *flag, int length)
 int hexa_print(va_list arg, flag_t *flag, int length)
 {
 	int i = 0, j = 0, digit = 0, k;
-	unsigned int num = va_arg(arg, unsigned int);
+	unsigned long int num = va_arg(arg, unsigned long int);
 
 	num = lengthConverter(num, length);
 	digit = hexa_convert(num, -1, 0);
@@ -92,7 +92,7 @@ int hexa_print(va_list arg, flag_t *flag, int length)
 int upHexa_print(va_list arg, flag_t *flag, int length)
 {
 	int i = 0, j = 0, digit = 0, k;
-	unsigned int num = va_arg(arg, unsigned int);
+	unsigned long int num = va_arg(arg, unsigned long int);
 
 	num = lengthConverter(num, length);
 	digit = hexa_convert(num, 1, 0);
